delivery_lysto/requirements.txt
flask>=3.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pymysql>=1.1.0
